all by declared type of object removed ObjectKey Objects use float64 for extra precision during trigo operations until dumped to OpenGL buffer in float32 Fixed sphere cutting issue in Z plane

+++ m3gl/var3gl.go
@@ -1,13 +1,13 @@
 package m3gl
 
 type SizeVar struct {
-	Min float32
-	Max float32
-	Val float32
+	Min float64
+	Max float64
+	Val float64
 }
 
-func (v SizeVar) getDelta() float32 {
-	return (v.Val - v.Min) / (10 * (v.Max - v.Min))
+func (v SizeVar) getDelta() float64 {
+	return (v.Val - v.Min) / (10.0 * (v.Max - v.Min))
 }
 
 func (v *SizeVar) check() {
